+++ mail/proton-mail-common/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - 2024-00-00
 
+## Changed
+
+- Update `MailUserContext::filter_converstions` to require a label id for context.
+
 ## [0.5.26] - 2024-06-21
 
 ### Changed